home_task_three/requirements.txt
pytest==7.4.2
selenium==4.13.0
webdriver-manager==4.0.1
python-dotenv==1.0.0